backend/requirements.txt
flask
flask-cors
pytesseract
Pillow
beautifulsoup4
requests 
gunicorn
opencv-python-headless